requirement.txt
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
cn-clip==1.5.1
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.0
datasets==2.14.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gensim==4.3.1
grpcio==1.56.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imgaug==0.4.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jstyleson==0.0.2
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.6
lmdb==1.3.0
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==2.8.8
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
openvino==2023.0.1
openvino-dev==2023.0.1
openvino-telemetry==2023.1.1
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.7.0.0
paddlepaddle==2.5.1
pandas==2.0.3
pdf2docx==0.5.6
Pillow==10.0.0
pip==23.2.1
premailer==3.10.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pymilvus==2.2.14
PyMuPDF==1.20.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.2.0
rarfile==4.0
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.12
termcolor==2.3.0
text2vec==1.2.2
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.5
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
visualdl==2.5.3
Werkzeug==2.3.6
wheel==0.38.4
xxhash==3.3.0
yarl==1.9.2
